+++ src/components/watchlistDetail/data/detailData.js
