+++ src/components/watchlistDetail/data/otherData.js
