+++ src/components/watchlistDetail/style/WatchListDetailMovieList.js
