+++ src/components/watchlistDetail/style/WatchListDetailStorage.js
