+++ src/components/watchlistEdit/style/EditStorage.js
